🐛 fix: 순환참조 에러 해결 위한 파일 분리 #334
